Foreign.C.Error.Safe

Description

Provides a variant of the Foreign.C.Error API which returns errors explicitly, instead of throwing exceptions.

Since: 0.3.5

Primitive handlers

eitherErrnoIf

Arguments

:: (a -> Bool)
Predicate to apply to the result value of the IO operation.

-> IO a
The IO operation to be executed.

-> IO (Either Errno a)

A variant of throwErrnoIf which returns Either instead of throwing an errno error.

eitherErrnoIfRetry

Arguments

:: (a -> Bool)
Predicate to apply to the result value of the IO operation.

-> IO a
The IO operation to be executed.

-> IO (Either Errno a)

A variant of throwErrnoIfRetry which returns Either instead of throwing an errno error.

eitherErrnoIfRetryMayBlock

Arguments

:: (a -> Bool)
Predicate to apply to the result value of the IO operation.

-> IO a
The IO operation to be executed.

-> IO b
Action to execute before retrying if an immediate retry would block.

-> IO (Either Errno a)

A variant of throwErrnoIfRetryMayBlock which returns Either instead of throwing an errno error.

Derived handlers

With predicate (-1 ==)

eitherErrnoIfMinus1 :: (Eq a, Num a) => IO a -> IO (Either Errno a)

eitherErrnoIfMinus1Retry :: (Eq a, Num a) => IO a -> IO (Either Errno a)

eitherErrnoIfMinus1RetryMayBlock :: (Eq a, Num a) => IO a -> IO b -> IO (Either Errno a)

With predicate (nullPtr ==)

eitherErrnoIfNull :: IO (Ptr a) -> IO (Either Errno (Ptr a))

eitherErrnoIfNullRetry :: IO (Ptr a) -> IO (Either Errno (Ptr a))

eitherErrnoIfNullRetryMayBlock :: IO (Ptr a) -> IO b -> IO (Either Errno (Ptr a))